# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是後端 API 文件",
        "title": "interchange API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/admin/users/{userID}/api-keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-APIKey"
                ],
                "summary": "取得用戶的 API Key 列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interchange_internal_dto.UserAPIKeyResponseDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-APIKey"
                ],
                "summary": "為指定用戶新增 API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "API Key 資訊",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.CreateUserAPIKeyDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UserAPIKeyResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{userID}/api-keys/{apiKeyID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-APIKey"
                ],
                "summary": "取得單一 API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UserAPIKeyResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-APIKey"
                ],
                "summary": "全量覆蓋 provider_access 欄位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "provider_access",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interchange_internal_dto.UpdateProviderAccessAllDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-APIKey"
                ],
                "summary": "刪除 API Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{userID}/api-keys/{apiKeyID}/provider/{provider}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin-APIKey"
                ],
                "summary": "局部更新 provider 欄位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的欄位map",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "interchange_internal_core.ApiScope": {
            "type": "string",
            "enum": [
                "/chat/completions",
                "/images/generations",
                "/images/variations",
                "/images/edits",
                "/audio/transcriptions",
                "/embeddings",
                "/models",
                "/mcp-server/*",
                "*"
            ],
            "x-enum-varnames": [
                "ApiScopeChatCompletions",
                "ApiScopeImagesGenerations",
                "ApiScopeImagesVariations",
                "ApiScopeImagesEdits",
                "ApiScopeAudioTranscriptions",
                "ApiScopeEmbeddingsGenerations",
                "ApiScopeGetModels",
                "ApiScopeMCPServer",
                "ApiScopeAll"
            ]
        },
        "interchange_internal_core.LimitPeriod": {
            "type": "string",
            "enum": [
                "daily",
                "weekly",
                "monthly",
                "yearly",
                "none"
            ],
            "x-enum-varnames": [
                "LimitPeriodDaily",
                "LimitPeriodWeekly",
                "LimitPeriodMonthly",
                "LimitPeriodYearly",
                "LimitPeriodNone"
            ]
        },
        "interchange_internal_core.ProviderName": {
            "type": "string",
            "enum": [
                "openai",
                "gemini",
                "grok",
                "custom"
            ],
            "x-enum-varnames": [
                "ProviderOpenAI",
                "ProviderGemini",
                "ProviderGrok",
                "ProviderCustom"
            ]
        },
...
        "interchange_internal_core.Status": {
            "type": "string",
            "enum": [
                "active",
                "blocked",
                "suspended",
                "expired",
                "revoked",
                "maintenance",
                "pending",
                "deleted"
            ],
            "x-enum-comments": {
                "StatusActive": "正常可用",
                "StatusBlocked": "被封鎖（例如濫用）",
                "StatusDeleted": "已刪除（軟刪除）",
                "StatusExpired": "已過期",
                "StatusMaintenance": "系統維護中（暫時停用）",
                "StatusPending": "尚未啟用（等待審核/激活）",
                "StatusRevoked": "被手動撤銷",
                "StatusSuspended": "暫停（違規調查中）"
            },
            "x-enum-varnames": [
                "StatusActive",
                "StatusBlocked",
                "StatusSuspended",
                "StatusExpired",
                "StatusRevoked",
                "StatusMaintenance",
                "StatusPending",
                "StatusDeleted"
            ]
        },
        "interchange_internal_dto.CreateUserAPIKeyDto": {
            "type": "object",
            "required": [
                "keyName",
                "providerAccess"
            ],
            "properties": {
                "keyName": {
                    "type": "string"
                },
                "keyValue": {
                    "type": "string"
                },
                "providerAccess": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/interchange_internal_dto.ProviderAccessDto"
                    }
                }
            }
        },
...
        "interchange_internal_dto.ProviderAccessDto": {
            "type": "object",
            "required": [
                "provider",
                "status"
            ],
            "properties": {
                "apiScopes": {
                    "description": "可選，API 權限範圍",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/interchange_internal_core.ApiScope"
                    }
                },
                "expireTime": {
                    "description": "API Key 過期時間",
                    "type": "string"
                },
                "lastResetAt": {
                    "description": "最後重置時間",
                    "type": "string"
                },
                "lastSeen": {
                    "description": "最後使用時間",
                    "type": "string"
                },
                "limitCount": {
                    "description": "限制次數",
                    "type": "integer"
                },
                "limitPeriod": {
                    "description": "daily, weekly, monthly...",
                    "allOf": [
                        {
                            "$ref": "#/definitions/interchange_internal_core.LimitPeriod"
                        }
                    ]
                },
                "provider": {
                    "description": "openai, gemini...",
                    "allOf": [
                        {
                            "$ref": "#/definitions/interchange_internal_core.ProviderName"
                        }
                    ]
                },
                "providerKey": {
                    "description": "可選，對應 Provider 的 API Key",
                    "type": "string"
                },
                "status": {
                    "description": "active, blocked...",
                    "allOf": [
                        {
                            "$ref": "#/definitions/interchange_internal_core.Status"
                        }
                    ]
                },
                "usedCount": {
                    "description": "已使用次數",
                    "type": "integer"
                }
            }
        },
        "interchange_internal_dto.UpdateProviderAccessAllDto": {
            "type": "object",
            "required": [
                "providerAccess"
            ],
            "properties": {
                "providerAccess": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/interchange_internal_dto.ProviderAccessDto"
                    }
                }
            }
        },
...
        "interchange_internal_dto.UserAPIKeyResponseDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "keyName": {
                    "type": "string"
                },
                "keyValue": {
                    "type": "string"
                },
                "providerAccess": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/interchange_internal_dto.ProviderAccessDto"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "請在欄位輸入 \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}